project structure refactoring
